Projeto finalizada

+++ src/components/MovieCard.jsx
@@ -1,13 +1,19 @@
-import { Link, useNavigate } from "react-router-dom"
+import { Link, useNavigate } from "react-router-dom";
 
 export default function MovieCard({ id, title, poster_path }) {
     const navigate = useNavigate()
 
-    const addToWatchLater = (movie) => {
+    const addToWatchLater = () => {
+        const movie = { id, title, poster_path } // Criando o objeto filme
         let watchLaterMovies = JSON.parse(localStorage.getItem('watchLaterMovies')) || []
-        watchLaterMovies.push(movie)
-        localStorage.setItem('watchLaterMovies', JSON.stringify(watchLaterMovies))
-        navigate('/mylist')
+        
+        // Verifica se o filme já não está na lista
+        const alreadyInList = watchLaterMovies.some(m => m.id === movie.id)
+        if (!alreadyInList) {
+            watchLaterMovies.push(movie)
+            localStorage.setItem('watchLaterMovies', JSON.stringify(watchLaterMovies))
+            navigate('/mylist') // Redireciona para a página 'mylist'
+        }
     }
 
     return (
@@ -19,7 +25,6 @@ export default function MovieCard({ id, title, poster_path }) {
             delay-100
             hover:scale-105 hover:shadow-sm hover:bg-gray-700 hover:rounded-md
         ">
-            
             <img 
                 src={`https://image.tmdb.org/t/p/w1280${poster_path}`} 
                 alt={title} 
@@ -27,7 +32,7 @@ export default function MovieCard({ id, title, poster_path }) {
             />
             <h2 className="text-xs text-center">{title}</h2>
             <Link to={`/movies/${id}`} className="text-blue-500 mt-1 text-xs">Saber mais</Link>
-            <button className="h-auto w-10 text-center text-xs bg-zinc-950 rounded-2xl" onClick={() => addToWatchLater(movie)}>Add</button>
+            <button className="h-auto w-10 text-center text-xs bg-zinc-950 rounded-2xl" onClick={addToWatchLater}>Add</button>
         </div>
     );
 }
